+++ src/app/components/statusi-tabela/statusi-tabela.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, Input, Host } from '@angular/core';
 import { Status } from 'src/app/models/Status';
 import { StatusiService } from 'src/app/services/statusi.service';
 import { AutomobiliAdmComponent } from '../automobili-adm/automobili-adm.component';
+import { AuthService } from 'src/app/services/auth.service';
 
 @Component({
   selector: 'app-statusi-tabela',
@@ -14,34 +15,45 @@ export class StatusiTabelaComponent implements OnInit {
   statusi: Status[] = [];
 
   constructor(private statusiService: StatusiService, 
-              @Host() private parent: AutomobiliAdmComponent) { }
+              @Host() private parent: AutomobiliAdmComponent, 
+              private authService: AuthService) { }
 
   ngOnInit(): void {
   }
 
   ukloniStatus(status: Status) {
-    if (confirm('Da li zelite da uklonite status?')) {
-      this.statusiService.deleteStatus(status.id).subscribe(data => {
-        if (data.status === 0) {
-          alert('Status je uklonjen iz baze podataka!');
-          this.parent.ngOnInit();
-        }
-        else {
-          alert('Doslo je do greske pri uklanjanju iz baze podataka!');
-        }
-      });
+    if(window.localStorage.getItem('ia-token') && this.authService.isLoggedIn() && (this.authService.getKorisnikDetails().isAdmin === 1)) {
+      if (confirm('Da li zelite da uklonite status?')) {
+        this.statusiService.deleteStatus(status.id).subscribe(data => {
+          if (data.status === 0) {
+            alert('Status je uklonjen iz baze podataka!');
+            this.parent.ngOnInit();
+          }
+          else {
+            alert('Doslo je do greske pri uklanjanju iz baze podataka!');
+          }
+        });
+      }
+    }
+    else {
+      alert('Nemate administratorska prava!');
     }
   }
 
   izmijeniStatus(status: Status) {
-    if (confirm('Da li zelite da izmijenite status?')) {
-      this.parent.odabranaIzmjUklanjOst = false;
-      this.parent.odabranUnosPrMoSt = true;
-      this.parent.proizvodjacVidljiv = false;
-      this.parent.modelVidljiv = false;
-
-      this.parent.odabraniStatus = status;
-      this.parent.dodajStatusBtn = false;
+    if (window.localStorage.getItem('ia-token') && this.authService.isLoggedIn() && (this.authService.getKorisnikDetails().isAdmin === 1)) {
+      if (confirm('Da li zelite da izmijenite status?')) {
+        this.parent.odabranaIzmjUklanjOst = false;
+        this.parent.odabranUnosPrMoSt = true;
+        this.parent.proizvodjacVidljiv = false;
+        this.parent.modelVidljiv = false;
+  
+        this.parent.odabraniStatus = status;
+        this.parent.dodajStatusBtn = false;
+      }
+    }
+    else {
+      alert('Nemate administratorska prava!');
     }
   }
